added validation and overall fixes , added @PreAuthorize , added confirmed field to comment , added content to post , added media controller and some changes overall

+++ src/main/java/com/pouryat/headless_cms/service/CommentServiceImpl.java
@@ -1,15 +1,22 @@
 package com.pouryat.headless_cms.service;
 
+import com.pouryat.headless_cms.auth.jwt.utils.JwtUtils;
 import com.pouryat.headless_cms.dto.CommentCreateDto;
 import com.pouryat.headless_cms.dto.CommentResponseDto;
 import com.pouryat.headless_cms.entity.Comment;
+import com.pouryat.headless_cms.entity.Role;
+import com.pouryat.headless_cms.entity.User;
 import com.pouryat.headless_cms.mapper.CommentMapper;
+import com.pouryat.headless_cms.model.RoleName;
 import com.pouryat.headless_cms.repository.CommentRepository;
 import com.pouryat.headless_cms.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -18,40 +25,76 @@ import java.util.stream.Collectors;
 public class CommentServiceImpl implements CommentService {
     private final CommentRepository commentRepository;
     private final PostRepository postRepository;
-    private final CommentMapper commentMapper;
 
     @Override
-    public CommentResponseDto createComment(CommentCreateDto dto) {
-        Comment comment = commentMapper.toEntity(dto);
-        comment.setPost(postRepository.findById(dto.getPostId())
-                .orElseThrow(() -> new RuntimeException("Post not found")));
-        comment.setAuthorName(SecurityContextHolder.getContext().getAuthentication().getName());
-        return commentMapper.toDTO(commentRepository.save(comment));
+    public List<CommentResponseDto> getAllComments(int page, int size) {
+        Pageable pageable = PageRequest.of(page, size);
+        return commentRepository.findAll(pageable).stream()
+                .map(CommentMapper::toDTO)
+                .collect(Collectors.toList());
     }
 
     @Override
     public CommentResponseDto getCommentById(Long id) {
-        return commentMapper.toDTO(commentRepository.findById(id)
+        return CommentMapper.toDTO(commentRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Comment not found")));
     }
 
     @Override
-    public CommentResponseDto updateComment(Long id, CommentCreateDto dto) {
+    public List<CommentResponseDto> getCommentByPostId(Long id, int page, int size) {
+        Pageable pageable = PageRequest.of(page, size);
+        return commentRepository.findAllByPostId(id, pageable).stream()
+                .map(CommentMapper::toDTO)
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public CommentResponseDto createComment(User user, CommentCreateDto dto) {
+        boolean status = false;
+
+        for (Role role : user.getRoles()) {
+            if (role.getRoleName().equals(RoleName.ROLE_ADMIN)) {
+                status = true;
+                break;
+            }
+        }
+        Comment comment = Comment.builder()
+                .createdAt(LocalDateTime.now())
+                .author(user)
+                .content(dto.getText())
+                .confirmed(status)
+                .post(postRepository.findById(dto.getPostId()).orElseThrow(() -> new RuntimeException("Post not found")))
+                .build();
+
+        return CommentMapper.toDTO(commentRepository.save(comment));
+    }
+
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
+    @Override
+    public CommentResponseDto updateCommentStatus(Long id, boolean confirmed) {
         Comment comment = commentRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Comment not found"));
+        comment.setConfirmed(true);
+        return CommentMapper.toDTO(commentRepository.save(comment));
+    }
+
+    @Override
+    public CommentResponseDto updateComment(User user, Long id, CommentCreateDto dto) {
+        Comment comment = commentRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Comment not found"));
+        JwtUtils.checkOwnership(comment.getAuthor().getId(), user.getId());
+
         comment.setContent(dto.getText());
-        return commentMapper.toDTO(commentRepository.save(comment));
+        comment.setAuthor(user);
+        return CommentMapper.toDTO(commentRepository.save(comment));
     }
 
     @Override
-    public void deleteComment(Long id) {
+    public void deleteComment(User user, Long id) {
+        Comment comment = commentRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Comment not found"));
+        JwtUtils.checkOwnership(comment.getAuthor().getId(), user.getId());
+
         commentRepository.deleteById(id);
-    }
-
-    @Override
-    public List<CommentResponseDto> getAllComments() {
-        return commentRepository.findAll().stream()
-                .map(commentMapper::toDTO)
-                .collect(Collectors.toList());
     }
 }
